+++ src/dash/Dashboard.jsx
@@ -7,7 +7,7 @@ import { Provider } from 'react-redux';
 import store from '../redux/store.js';
 import AboutUs from '../home/AboutUs.jsx';
 import {Router, Route, browserHistory} from 'react-router';
-import { updateUser } from '../redux/actions.js';
+import { updateUser } from '../redux/actions/user_actions.js';
 import { connect } from 'react-redux';
 import axios from 'axios';
