+++ src/redux/actions/student_view_actions.js
@@ -1,10 +1,3 @@
-export function updateUser (user) {
-    return {
-        type: 'UPDATE_IN_USER',
-        payload: user
-    };
-}
-
 export function changeOnlineTutors(tutors) {
   return {
     type: 'CHANGE_ONLINE_TUTORS',
