+++ src/redux/actions/user_actions.js
@@ -1,0 +1,6 @@
+export function updateUser (user) {
+    return {
+        type: 'UPDATE_IN_USER',
+        payload: user
+    };
+}
